timerUnit.c und timerUnit.h: timer_setCounter() refaktorisiert: Werteintervall vom Argument erweitert, Benutzer-Schnittstelle statt Hardware-Schnittstelle benutzt

+++ 01 Hardware/01 Lidar/03 Software/01 source/Testen mit USVData/Testen mit USVData/timerUnit.h
@@ -113,7 +113,7 @@ enum prescaler{
 
 extern uint8_t timerInit(uint8_t rezConfig, uint8_t resolution);
 extern void timer_setState(uint8_t state);
-extern void timer_setCounter(uint16_t value);
+extern void timer_setCounter(int32_t value);
 extern const uint16_t* timer_getCounter();
 extern void timer_stopWatch(uint16_t val);
 extern void waitCycle(uint32_t cycle);
